% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imtopix.R
\name{imtopix}
\alias{imtopix}
\title{Convert Image to Pseudo Pixel Art}
\usage{
imtopix(
  im,
  blockSize = 8,
  pal = c("appleii", "atari2600", "commodore64", "contra", "gameboy", "grayscale",
    "hyrule", "intellivision", "kungfu", "mario", "nes", "sega", "tetris")
)
}
\arguments{
\item{im}{an image; a cimg object, normally loaded by
\code{imager::load.image()}.}

\item{blockSize}{numeric; the size of single pixel or block.}

\item{pal}{character; colour palette used in matching.}
}
\description{
Convert image to pseudo pixel art by reducing the resolution of the image and
matching the colour with a colour palette.
}
\section{Copyright Notice}{

This algorithm is a reduced version of a python script written by
Nathan Harper. The intention of this package is to provide an R
API of this algorithm. The author of this R package do not own
the copyright of this algorithm. Please check the initial source of this
algorithm \url{https://github.com/nathanharper/phixelgator} and Nathan
Harper's Github page \url{https://github.com/nathanharper} for more details.
}

\section{Performance Issue}{

R takes noticeable time to run this algorithm. If you would like to save your
time, please use Nathan Harper's original script written in Python
\url{https://github.com/nathanharper/phixelgator}.
}

\examples{
im <- sample_im
imtopix(im, pal = "appleii")

}
