% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pal_collection}
\alias{pal_collection}
\title{13 Colour Palettes}
\format{
An object of class \code{list} of length 13.
}
\source{
\url{https://github.com/nathanharper/phixelgator/tree/master/palettes/rgb}
}
\usage{
pal_collection
}
\description{
A list containing 13 colour palettes representing in RGB format.
}
\keyword{datasets}
