% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_im}
\alias{sample_im}
\title{A sample image}
\format{
An object of class \code{cimg} (inherits from \code{imager_array}, \code{numeric}) of dimension 1080 x 1440 x 1 x 3.
}
\usage{
sample_im
}
\description{
A sample image.
}
\keyword{datasets}
